migration change type to number

+++ prisma/migrations/20220525190009_init/migration.sql
@@ -60,7 +60,7 @@ CREATE TABLE "Raquete" (
     "corte" TEXT,
     "dataObservacao" TEXT,
     "exported" BOOLEAN NOT NULL,
-    "nextVersion" BIGINT NOT NULL,
+    "nextVersion" INTEGER NOT NULL,
     "estacaoId" INTEGER,
 
     CONSTRAINT "Raquete_pkey" PRIMARY KEY ("id")
